Beginning development from scratch
After deciding to resume development of this personal project in my free time, I realised that I had poorly planned the development timeline of the project. Thus, the decision was made to scrap all existing work and start the project from scratch. Add files are just some simple foundational code for the program.