Full Implementation of First Algorithm
This commit encompasses the completion and full implementation of the first encryption algorithm for this personal program/project. The implemented algorithm is a simple substitution algorithm.

I also created some makefiles for the users convenience. While the commands themselves work, I myself have not been able to test the makefiles fully as I am encountering issues installing and getting make working on my device.

There was also a small updated to the README.md file.